LayoutTests/media/modern-media-controls/macos-media-controls/macos-media-controls.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSMediaControls</code> constructor.");

const mediaControls = new MacOSMediaControls;

shouldBeEqualToString("mediaControls.element.localName", "div");
shouldBeEqualToString("mediaControls.element.className", "media-controls mac");
shouldBe("mediaControls.layoutTraits", "LayoutTraits.macOS");
shouldBeTrue("mediaControls.muteButton instanceof MuteButton");
shouldBeTrue("mediaControls.tracksButton instanceof TracksButton");
shouldBeTrue("mediaControls.volumeSlider instanceof VolumeSlider");

debug("");
const mediaControlsWithArguments = new MacOSMediaControls({ width: 600, height: 300 });
shouldBe("mediaControlsWithArguments.width", "600");
shouldBe("mediaControlsWithArguments.height", "300");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
